+++ OnlySnarf/classes/webdriver/post.py
@@ -9,7 +9,6 @@ from selenium.common.exceptions import TimeoutException
 
 from .expiration import expiration as EXPIRES
 from .errors import error_checker
-from .message import message_clear
 from .poll import poll as POLL
 from .schedule import schedule as SCHEDULE
 from .upload import upload_files
@@ -46,7 +45,7 @@ def post(browser, post_object):
     if not post_object:
         logging.debug("skipping empty post")
         return True
-    message_clear(browser)
+    clear_button(browser)
     #################### Formatted Text ####################
     logging.info("====================")
     logging.info("Posting:")
@@ -77,7 +76,7 @@ def post(browser, post_object):
     except Exception as e:
         logging.debug(e)
         logging.error("failed to send post!")
-    message_clear(browser)
+    clear_button(browser)
     return True
 
 def enter_text(browser, text):
@@ -100,6 +99,8 @@ def enter_text(browser, text):
         """
 
         try:
+            from .message import clear_text
+            clear_text(browser)
             logging.debug("entering text: "+text)
             element = browser.find_element(By.ID, "new_post_text_input")
             action = ActionChains(browser)
@@ -107,7 +108,7 @@ def enter_text(browser, text):
             action.click(on_element=element)
             action.double_click()
             action.click_and_hold()
-            action.send_keys(Keys.CLEAR)
+            # action.send_keys(Keys.CLEAR)
             action.send_keys(str(text))
             action.perform()
             logging.debug("successfully entered text!")
@@ -180,3 +181,20 @@ def open_more_options(browser):
         error_checker(e)
     
     raise Exception("unable to locate 'More Options' element")
+
+def clear_button(browser, retry=False):
+    try:
+        find_element_to_click(browser, "button", by=By.TAG_NAME, text="Clear").click()
+        logging.debug("successfully clicked clear button!")
+        return True
+    except Exception as e:
+        if not retry:
+            go_to_home(browser, force=True)
+            action = ActionChains(browser)
+            action.move_to_element(browser.find_element(By.ID, "new_post_text_input"))
+            action.click(on_element=browser.find_element(By.ID, "new_post_text_input"))
+            action.perform()
+            time.sleep(0.5) # needs to load: TODO: possibly add wait
+            return clear_button(browser, retry=True)
+        logging.debug("unable to click clear button!")
+    return False