+++ tests/snarf/test_message.py
@@ -24,21 +24,21 @@ class TestSnarf(unittest.TestCase):
     def test_message(self):
         assert self.message.send(), "unable to send basic message"
 
-    def test_message_files_local(self):
-        CONFIG["input"] = ["/home/skeetzo/Projects/onlysnarf/public/images/shnarf.jpg", "/home/skeetzo/Projects/onlysnarf/public/images/snarf.jpg"]
-        self.message = Message.create_message({**CONFIG})
-        assert self.message.send(), "unable to upload message files - local"
+    # def test_message_files_local(self):
+    #     CONFIG["input"] = ["/home/skeetzo/Projects/onlysnarf/public/images/shnarf.jpg", "/home/skeetzo/Projects/onlysnarf/public/images/snarf.jpg"]
+    #     self.message = Message.create_message({**CONFIG})
+    #     assert self.message.send(), "unable to upload message files - local"
 
-    def test_message_files_remote(self):
-        CONFIG["input"] = ["https://github.com/skeetzo/onlysnarf/blob/master/public/images/shnarf.jpg?raw=true", "https://github.com/skeetzo/onlysnarf/blob/master/public/images/snarf.jpg?raw=true"]
-        self.message = Message.create_message({**CONFIG})
-        assert self.message.send(), "unable to upload message files - remote"
+    # def test_message_files_remote(self):
+    #     CONFIG["input"] = ["https://github.com/skeetzo/onlysnarf/blob/master/public/images/shnarf.jpg?raw=true", "https://github.com/skeetzo/onlysnarf/blob/master/public/images/snarf.jpg?raw=true"]
+    #     self.message = Message.create_message({**CONFIG})
+    #     assert self.message.send(), "unable to upload message files - remote"
 
-    def test_message_price(self):
-        CONFIG["input"] = ["/home/skeetzo/Projects/onlysnarf/public/images/shnarf.jpg"]
-        CONFIG["price"] = DEFAULT.PRICE_MINIMUM
-        self.message = Message.create_message({**CONFIG})
-        assert self.message.send(), "unable to set message price"
+    # def test_message_price(self):
+    #     CONFIG["input"] = ["/home/skeetzo/Projects/onlysnarf/public/images/shnarf.jpg"]
+    #     CONFIG["price"] = DEFAULT.PRICE_MINIMUM
+    #     self.message = Message.create_message({**CONFIG})
+    #     assert self.message.send(), "unable to set message price"
 
 ############################################################################################